amez/requirements.txt
anyascii==0.3.1
asgiref==3.6.0
beautifulsoup4==4.11.1
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.5
django-filter==22.1
django-modelcluster==6.0
django-permissionedforms==0.1
django-taggit==3.1.0
django-treebeard==4.6.0
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
html5lib==1.1
idna==3.4
l18n==2021.3
openpyxl==3.0.10
Pillow==9.4.0
pytz==2022.7.1
requests==2.28.2
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
telepath==0.3
tzdata==2022.7
urllib3==1.26.14
wagtail==4.1.1
webencodings==0.5.1
Willow==1.4.1
